:recycle: Extract other file for weekday
